rename gtm to analytics, fix tests
