+++ __mocks__/@redux-beacon/ensure.js
@@ -1,0 +1,3 @@
+const ensureMock = jest.fn((arg1, arg2) => jest.fn(() => arg2));
+
+export default ensureMock;
