+++ src/helpers/analytics/__mocks__/eventsDefinitions.js
