+++ src/helpers/analytics/eventsDefinitions.js
