+++ src/helpers/analytics/tests/eventsDefinitions.test.js
